fisherconn_1.0/requirements.txt
coreapi==2.2.4
coreschema==0.0.4
dj-database-url==0.4.1
Django==1.10.4
django-ajax-search==1.5.1
django-allauth==0.29.0
django-bootstrap-form==3.2.1
django-crispy-forms==1.6.1
django-filter==1.0.1
django-haystack==2.5.1
django-social-auth==0.7.28
djangorestframework==3.5.4
Djapian==2.3
elasticsearch==5.1.0
elasticsearch-django==0.5.2
elasticsearch-dsl==5.1.0
gunicorn==19.6.0
httplib2==0.9.2
ipython==4.1.2
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==0.23
oauth2==1.9.0.post1
oauthlib==2.0.1
path.py==8.1.2
Pillow==3.4.2
psycopg2==2.6.2
Pygments==2.2.0
pystache==0.5.4
python-dateutil==2.6.0
python-openid==2.2.5
requests==2.12.4
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
static==1.1.1
uritemplate==3.0.0
urllib3==1.20
virtualenv==14.0.6
whitenoise==3.2.2
Whoosh==2.7.4
